Finalize 1 (#59)
* clean debugging logs/prints

* allow headers in request using @

* working SET and LOG

* fix dvar access

* rename to DataEvaluator, temporary(nonlocal) fix for loop, allow .AQLKEY and .AQLVALUE to access generic key and value from JSON

* remove singlequote string (JSON doesnt use it), simplify params and WITH

* use real endpoints for tests and fix antlrtests

+++ src/controller/helper/StringParser.java
@@ -13,7 +13,7 @@ public class StringParser {
     public static Object parseValue(String text) {
         try {
             if (text.matches("-?\\d+")) {
-                return Integer.parseInt(text);
+                return Double.parseDouble(text);
             } else if (text.matches("-?\\d+\\.\\d+")) {
                 return Double.parseDouble(text);
             }
